Progress service to update graph after task completion & refactor content.board logic
New progress.service, which can trigger a graph update (here we could also always update the progress at the bottom right).
Refactor content.board component, so that when the PDF/video view is closed, it is also checked whether 100% has been reached for the content. As soon as 100% has been reached after PDF/Video or a task, an update of the graph is triggered with the progress.service.

Added some "awaits". This way we get the result AFTER the level has been updated. Now we can refresh the graph and get the new levels. To do this we use the method in the new progress.service after a task has been completed or a video/pdf has been "understood".
- to questionContentElementDone.
- added await to checkUserConceptLevelAward so that it does not return before the level has been updated in the db (same reason as above)
- added await to this.updateUserLevel (same reason as above)
- added await to this.contentService.toggleCheckmark (same reason as above)

Issue:
- When user removes understood pdf or video, the progress in content gets removed correctly, but we dont have a way to remove the level in graph yet.