UserSubject added and seed updated
New relationship between users and subject to define their connections, like if they are registered for SL or maybe some students will be teaching assistants for a specific subject. Had to rework auth a bit for this. Also moved seedFiles to its own seed. Removed the "modules" module because it is not used and we want to use subjects instead later.