#153 prevent multiple cas accounts with big/small "g"
After logging in via cas, students receive an account based on the g number. Since they could log in to cas with a lower or upper case g, this led to two possible accounts. This was simply fixed by setting the email (and thus the g-number contained therein) in lowercase when creating and searching for accounts.

Problem:
If users can register later with their own e-mail, this may only be written in lowercase.
 Currently this is not a problem, as all logins by mail are only possible through the accounts created in the seed (these are already lowercase).